feat!: abstract the routes
