+++ packages/multiply-backend/src/routes/__init__.py
@@ -1,0 +1,5 @@
+from flask import Flask
+from src.routes.UserRoutes import userRoutes
+
+def registerRoutes(app: Flask):
+    app.register_blueprint(userRoutes)
